StarRating added

+++ src/main/java/wineverse/com/mk/Wineverse/Model/Review.java
@@ -14,14 +14,14 @@ public class Review {
     // author == user
     @ManyToOne
     private User author;
-    private Float rating;
+    private Integer rating;
     private String content;
     private LocalDate date;
 
     public Review() {
     }
 
-    public Review(User author, Float rating, String content, LocalDate date) {
+    public Review(User author, Integer rating, String content, LocalDate date) {
         //TODO for testing uncomment this
 //        Id = (long) (Math.random() * 1000);
         this.author = author;
